fix(lore-0172): address Copilot review on PR #143
- `EventItem.amount` is now `i64` to match the BIGINT column in
  `soroban_events_appearances` (was silently clamping to `i32::MAX` on
  out-of-range values). `expand_events` drops the `i32::try_from` clamp.
- Restore safe-retry pagination on archive outages: when no row in the
  current page expands successfully (`last_consecutive_idx == None`),
  `next_cursor` echoes the incoming cursor instead of returning `null`.
  Returning `null` would have caused clients to restart from the top
  on transient S3 outages, producing duplicates / infinite loops.

`InvocationItem.amount` stays `i32` — the underlying
`soroban_invocations_appearances.amount` column is `INTEGER`, not `BIGINT`.

+++ crates/api/src/contracts/dto.rs
@@ -54,7 +54,7 @@ pub struct EventItem {
     pub ledger_sequence: i64,
     pub transaction_id: i64,
     pub successful: bool,
-    pub amount: i32,
+    pub amount: i64,
     pub created_at: DateTime<Utc>,
     pub event_type: String,
     pub topics: Vec<serde_json::Value>,
